Create Day 51.py
It's an interactive to-do-list. It's possible to Add, remove and view. It saves in a file (When using for the first time, set the "file loading code" to be commented - it's just add a hashtag, aka, #, in the beginning of each line). After that, you can 'uncomment' it, taking of the hashtags, for the file is already saved.